Mise à jour de 'src/app/core_logic/responsable.h'

+++ src/app/core_logic/responsable.h
@@ -1,3 +1,3 @@
 
 
-int chargementArticles(int tRef[], float tPoids[], float tVol[], float tPrix[], float tPhysique);
+int chargementArticles(int tRef[], float tPoids[], float tVol[], float tPrix[], int tPhysique);
